+++ .netlify/functions-internal/___netlify-server-handler/.next/server/server-reference-manifest.js
@@ -1,0 +1,1 @@
+self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"dr0fWYRsx6esUVtaX58BbQDxjCN7/BvSusRknpyKlaI=\"\n}"